fix:  exception handeling

+++ src/main/java/com/shereen/catalog/controller/ItemController.java
@@ -22,75 +22,86 @@ import com.shereen.catalog.service.ItemService;
 @Controller
 @RequestMapping()
 public class ItemController {
-	
+
 	@Autowired
 	ItemService itemService;
-	
+
 	@Autowired
 	CatalogService catalogService;
-	
+
 	@GetMapping("{cataId}/items")
 	public String getAllItems(Model model, @PathVariable long cataId) {
-		Catalog  cata = catalogService.getCatalogById(cataId);
-		model.addAttribute("items", itemService.getAllItemsByCatalog( cata));
+		Optional<Catalog> cata = catalogService.getCatalogById(cataId);
+		if (!cata.isPresent())
+			throw new ResourceNotFoundException();
+		model.addAttribute("items", itemService.getAllItemsByCatalog(cata.get()));
 		model.addAttribute("catalog", cata);
 		return "item-list";
 	}
-	
+
 	@GetMapping("/JSON/{cataId}/items")
 	@ResponseBody()
-	public List<Item> getAllItemsJson(@PathVariable long cataId){
-		Catalog  cata = catalogService.getCatalogById(cataId);
-		return itemService.getAllItemsByCatalog( cata);
+	public List<Item> getAllItemsJson(@PathVariable long cataId) {
+		Optional<Catalog> cata = catalogService.getCatalogById(cataId);
+		if (!cata.isPresent())
+			throw new ResourceNotFoundException();
+		return itemService.getAllItemsByCatalog(cata.get());
 	}
-	
+
 	@GetMapping("{cataId}/items/{id}")
 	public String getItemByGuid(@PathVariable Long id, Model model, @PathVariable Long cataId) {
-		Catalog catalog = catalogService.getCatalogById(cataId);
-		Optional<Item> item = itemService.getItemByItemIdAndCatalog(id, catalog);
-		
-		if (!item.isPresent()) throw new ResourceNotFoundException();
-		
-		model.addAttribute("item", item.get() );
+		Optional<Catalog> catalog = catalogService.getCatalogById(cataId);
+		Optional<Item> item = itemService.getItemByItemIdAndCatalog(id, catalog.get());
+
+		if (!item.isPresent())
+			throw new ResourceNotFoundException();
+
+		model.addAttribute("item", item.get());
 		model.addAttribute("catalog", catalog);
 		return "item-view";
 	}
 
 	@GetMapping("{cataId}/items/new")
-	public String getItemForm(@PathVariable Long cataId,Model model) {
+	public String getItemForm(@PathVariable Long cataId, Model model) {
 		model.addAttribute("catalog", catalogService.getCatalogById(cataId));
 		return "item-new";
 	}
-	
+
 	@PostMapping("{cataId}/items/new")
-	public String addItem(@ModelAttribute Item item,@PathVariable Long cataId, Model model) {
-		System.out.println(item);
-		item.setCatalog(catalogService.getCatalogById(cataId));
+	public String addItem(@ModelAttribute Item item, @PathVariable Long cataId, Model model) {
+		Optional<Catalog> catalog = catalogService.getCatalogById(cataId);
+		if (!catalog.isPresent())
+			throw new ResourceNotFoundException();
+		item.setCatalog(catalog.get());
 		itemService.addItem(item);
-		
+
 		model.addAttribute("items", itemService.getAllItemsByCatalog(item.getCatalog()));
 		return ("redirect:/{cataId}/items");
 	}
 
 	@GetMapping("{cataId}/items/edit/{id}")
-	public String geteditItemForm( @PathVariable Long id, Model model) {
+	public String geteditItemForm(@PathVariable Long id, Model model) {
 		model.addAttribute("item", itemService.getItemById(id));
 		return ("item-edit");
 
 	}
-	
-	
+
 	@PostMapping("{cataId}/items/edit/{id}")
 	public String editItem(@PathVariable Long cataId, @PathVariable Long id, @ModelAttribute Item item, Model model) {
-		item.setCatalog(catalogService.getCatalogById(cataId));
+		if (!itemService.getItemById(id).isPresent())
+			throw new ResourceNotFoundException();
+		Optional<Catalog> catalog = catalogService.getCatalogById(cataId);
+		if (!catalog.isPresent())
+			throw new ResourceNotFoundException();
+		item.setCatalog(catalog.get());
 		itemService.updateItem(item);
 		model.addAttribute("items", itemService.getAllItemsByCatalog(item.getCatalog()));
 		return ("redirect:/{cataId}/items");
 
 	}
-	
+
 	@GetMapping("{cataId}/items/delete/{id}")
-	public String getDeleteItemForm( @PathVariable Long id, Model model) {
+	public String getDeleteItemForm(@PathVariable Long id, Model model) {
 		model.addAttribute("item", itemService.getItemById(id));
 		return ("item-delete");
 
@@ -98,12 +109,15 @@ public class ItemController {
 
 	@PostMapping("{cataId}/items/delete/{id}")
 	public String deleteItem(@PathVariable Long cataId, @PathVariable Long id, Model model) {
-		Catalog cata= catalogService.getCatalogById(cataId);
+		if (!itemService.getItemById(id).isPresent())
+			throw new ResourceNotFoundException();
+		Optional<Catalog> catalog = catalogService.getCatalogById(cataId);
+		if (!catalog.isPresent())
+			throw new ResourceNotFoundException();
 		itemService.deleteItem(id);
-		model.addAttribute("items", itemService.getAllItemsByCatalog(cata));
+		model.addAttribute("items", itemService.getAllItemsByCatalog(catalog.get()));
 		return ("redirect:/{cataId}/items");
 
 	}
-	
 
 }
